Delete internal clause package, move them to query.clause package
