--- thingif/src/main/java/com/kii/thingif/internal/clause/Clause.java
@@ -1,9 +1,0 @@
-package com.kii.thingif.internal.clause;
-
-import android.os.Parcelable;
-
-import org.json.JSONObject;
-
-public interface Clause extends Parcelable {
-    JSONObject toJSONObject();
-}
