+++ thingif/src/main/java/com/kii/thingif/query/clause/Clause.java
@@ -1,7 +1,9 @@
 package com.kii.thingif.query.clause;
 
+import android.os.Parcelable;
+
 import org.json.JSONObject;
 
-public interface Clause extends com.kii.thingif.internal.clause.Clause {
+public interface Clause extends Parcelable {
     JSONObject toJSONObject();
 }
