+++ thingif/src/main/java/com/kii/thingif/trigger/clause/Clause.java
@@ -1,4 +1,4 @@
 package com.kii.thingif.trigger.clause;
 
-public interface Clause extends com.kii.thingif.internal.clause.Clause {
+public interface Clause extends com.kii.thingif.query.clause.Clause {
 }
